/question CRUD

+++ src/repositories/questions.js
@@ -1,6 +1,6 @@
 import fs from 'fs/promises';
 
-import {QuestionInputModel} from '../models/question_data.js';
+import {QuestionData} from '../models/question_data.js';
 
 /**
  * Path to the Users database.
@@ -21,7 +21,7 @@ export const getNextQuestionID = async (dbPathOverride=questionsDBPath) => {
 
 /**
  * Create a new question in the database.
- * @param {QuestionInputModel} questionInputData - The input data of the
+ * @param {QuestionData} question - The input data of the
  * question to be created.
  * @param {string} dbPathOverride - Override of the path to the
  * database to be used. Defaults to the default defined path questionsDBPath.
@@ -29,16 +29,84 @@ export const getNextQuestionID = async (dbPathOverride=questionsDBPath) => {
  * question was successfully created.
  */
 export const createQuestion = async (
-    questionInputData,
+    question,
     dbPathOverride=questionsDBPath,
 ) => {
-  questionInputData = QuestionInputModel.fromJSON(questionInputData.toJSON());
   const questions = await getDatabase(dbPathOverride);
-  const question = questionInputData.question;
+  if (question.title.length === 0) {
+    throw new Error('Question must have a title.');
+  }
+  if (questions[question.question_id]) {
+    throw new Error(`Question with ID ${question.question_id} already exists.`);
+  }
   questions[question.question_id] = question.toJSON();
   fs.writeFile(dbPathOverride, JSON.stringify(questions));
   return true;
 };
+
+/**
+ * Fetch a question from the database.
+ * @param {string} questionID - The input data of the
+ * question to be created.
+ * @param {string} dbPathOverride - Override of the path to the
+ * database to be used. Defaults to the default defined path questionsDBPath.
+ * @return {Promise<QuestionData>} A promise that resolves to the question data.
+ */
+export const getQuestion = async (
+    questionID,
+    dbPathOverride=questionsDBPath,
+) => {
+  const questions = await getDatabase(dbPathOverride);
+  const qdJSON = questions[questionID];
+  return qdJSON ? QuestionData.fromJSON(qdJSON) : undefined;
+};
+
+/**
+ * Update a new question in the database.
+ * @param {string} questionID - The ID of the question to be updated.
+ * @param {QuestionData} question - The input data of the
+ * question to be updated. question.question_id is ignored
+ * @param {string} dbPathOverride - Override of the path to the
+ * database to be used. Defaults to the default defined path questionsDBPath.
+ * @return {Promise<boolean>} A promise that resolves to whether the
+ * question was successfully updated.
+ */
+export const updateQuestion = async (
+    questionID,
+    question,
+    dbPathOverride=questionsDBPath,
+) => {
+  const questions = await getDatabase(dbPathOverride);
+  if (question.title.length === 0) {
+    throw new Error('Question must have a title.');
+  }
+  if (!questions[questionID]) {
+    throw new Error(`Question with ID ${questionID} doesn't exist.`);
+  }
+  questions[questionID] = {
+    'question-ID': questionID,
+    ...question.toJSON(),
+  };
+  fs.writeFile(dbPathOverride, JSON.stringify(questions));
+  return true;
+};
+
+/**
+ * Deletes a question completely from the database if they exist.
+ * @param {string} questionID - ID of the user to be deleted.
+ * @param {string} dbPathOverride - Override of the path to the
+ * database to be used. Defaults to the default defined path usersDBPath.
+ * @return {Promise<void>} A promise that resolves when the user is deleted.
+ */
+export const deleteQuestion = async (
+    questionID,
+    dbPathOverride=usersDBPath,
+) => {
+  const questions = await getDatabase(dbPathOverride);
+  delete questions[questionID];
+  fs.writeFile(dbPathOverride, JSON.stringify(users));
+};
+
 
 /**
  * Returns a database object as an asynchronous promise.
